Seperate saltnado_websockets from the rest of the implementation.
