+++ doc/ref/netapi/all/salt.netapi.rest_tornado.rst
@@ -4,4 +4,6 @@ rest_tornado
 
 .. automodule:: salt.netapi.rest_tornado.saltnado
 
+.. automodule:: salt.netapi.rest_tornado.saltnado_websockets
+
 .. ............................................................................
